nova-compute: reduce default storage requirements
We have been allocating 80GB to root-disk for nova-compute and then a
further 50GB to the ephemeral storage anytime we used vault, which is
most of the time since we need vault for OVN. In those cases, the larger
80GB root-disk was never used.

Change the 50GB ephemeral storage to always be used regardless of vault,
and reduce the nova-compute root-disk to the normal size since it won't
be used.

Additionally adds a requirement for cores=2 as previously the flavors
available for a larger root-disk usually resulted in a 2-core machine.

Closes: #229